requirements-lock.txt
# requirements-lock.txt — Fully pinned pip dependencies
# Compatible with environment.yml (of3ds conda env, Python 3.12, CUDA 12.6 toolkit)
#
# IMPORTANT: Install ONLY inside the of3ds conda env after conda deps resolve.
#   conda activate of3ds && pip install --no-deps -r requirements-lock.txt
#
# Torch index: cu126 (forward-compatible with CUDA 12.8 drivers)
--extra-index-url https://download.pytorch.org/whl/cu126

# ── PyTorch core ─────────────────────────────────────────────────
torch==2.6.0+cu126
torchvision==0.21.0+cu126
torchaudio==2.6.0+cu126

# ── DeepSpeed ────────────────────────────────────────────────────
deepspeed==0.16.2

# ── PyTorch ecosystem ────────────────────────────────────────────
pytorch-lightning==2.5.0.post0
torchmetrics==1.6.1
lightning-utilities==0.11.9

# ── OpenFold3 core dependencies ──────────────────────────────────
biopython==1.85
dm-tree==0.1.8
ml-collections==1.0.0
modelcif==1.0

# ── Scientific / numerical ───────────────────────────────────────
# numpy and scipy are conda-managed; listed here as bounds check only
# numpy==1.26.4  (conda)
# scipy==1.14.1  (conda)
pandas==2.2.3
scikit-learn==1.6.1

# ── Data / IO ────────────────────────────────────────────────────
h5py==3.12.1
datasets==3.2.0
fsspec==2024.12.0
s3fs==2024.12.0
aiohttp==3.11.11
aiobotocore==2.15.2
botocore==1.35.76

# ── Config / CLI ─────────────────────────────────────────────────
pyyaml==6.0.2
pydantic==2.10.4
pydantic-core==2.27.2
click==8.1.8
typer==0.15.1

# ── HTTP / networking ────────────────────────────────────────────
requests==2.32.3
urllib3==2.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10

# ── Typing / compat ─────────────────────────────────────────────
typing_extensions==4.12.2
annotated-types==0.7.0
packaging==24.2
setuptools==75.8.0

# ── Logging / output ────────────────────────────────────────────
rich==13.9.4
tqdm==4.67.1
tabulate==0.9.0

# ── Jinja (DeepSpeed config templates) ───────────────────────────
jinja2==3.1.5
markupsafe==3.0.2

# ── Triton (PyTorch 2.6 dependency) ─────────────────────────────
triton==3.2.0

# ── Numba (optional; for accelerated kernels) ────────────────────
numba==0.61.0
llvmlite==0.44.0

# ── Hugging Face (model/data downloads) ──────────────────────────
huggingface-hub==0.27.1
safetensors==0.5.2
tokenizers==0.21.0
filelock==3.16.1

# ── Misc transitive deps (pinned for reproducibility) ────────────
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2024.2
regex==2024.11.6
sympy==1.13.1
mpmath==1.3.0
networkx==3.4.2
pillow==11.1.0
